changed logic a bit, schedules moved from utils to schedules.py; cos_wsd is testing
